fix(V2): switch to BFloat16 AMP + fix attention/residual in MaxViT/CoAtNet + update overseer report
Root cause: GradScaler+FP16 overflow in seg decoder backward path causes NaN
gradients in ALL 15 backbones (not just transformers). FP16 max=65504 is
insufficient for amplified gradients flowing through hook-captured features.

Core fix - BFloat16 (6 files):
- config.py: AMP_DTYPE = torch.bfloat16 (8-bit exponent, same range as FP32)
- train_v2_backbone.py: autocast(dtype=AMP_DTYPE), GradScaler disabled for BF16,
  removed prev_scaler/is_heavy_transformer params, simplified _run_phase
- oof_generator.py: BF16 autocast + disabled scaler
- stage1_individual_v2.py: BF16 autocast in inference
- smoke_training_pipeline.py: BF16 autocast + disabled scaler

Architecture fix - Base_backbones.py:
- MultiHeadSelfAttention: pre-multiply scale (q*scale before matmul)
- MultiHeadSelfAttention: FP32 attention matmul (cast Q,K to float32)
- CustomMaxViT: remove double residual in stage2_attn + stage3_transformer
- CustomCoAtNet: remove double residual in stage2_attn + stage3_transformer

Verification: 0 NaN grads for MaxViT+CoAtNet in both cls+seg backward with BF16.
All V2 artifacts cleaned for full Run 4 rerun.

Overseer report updated with Run 3 results, BF16 fix details, Run 4 pending.

+++ V2_segmentation/ensemble_v2/stage1_individual_v2.py
@@ -128,7 +128,7 @@ class Stage1IndividualV2:
         with torch.no_grad():
             for images, labels in loader:
                 images = images.to(DEVICE)
-                with torch.cuda.amp.autocast():  # type: ignore[attr-defined]
+                with torch.cuda.amp.autocast(dtype=torch.bfloat16):  # type: ignore[attr-defined]
                     outputs = model(images)
 
                 # Classification probabilities
